+++ sct_sentinel1_reader/src/sct_sentinel1_reader/corrections/__init__.py
@@ -5,3 +5,12 @@
 Sentinel-1 ALE Corrections
 --------------------------
 """
+
+ALE_CORRECTIONS_FIELDS = {
+    "etad_range": "ext_range_correction_[m]",
+    "etad_azimuth": "ext_azimuth_correction_[m]",
+    "rng_doppler_shift": "doppler_shift_range_correction_[m]",
+    "az_fmrate_shift": "fm_rate_shift_azimuth_correction_[m]",
+    "az_instrument_timing": "instrument_timing_azimuth_correction_[m]",
+    "az_bistatic_delay": "bistatic_delay_azimuth_correction_[m]",
+}